# Converted fasttable decoder to use __attribute__((musttail)) when it is available