The fasttable decoder relies on the compiler generating true tail calls when dispatching from one field to the next. Without tail calls, the decoder will use stack space proportional to the number of fields parsed, which will quickly lead to stack overflow.

Prior to this PR, we relied on the fact that Clang and GCC generally do generate tail calls when optimizations are enabled. However this is not a guarantee, and it definitely won't happen in debug modes. This makes the fasttable decoder too fragile to truly roll out into production environments.

However, thanks to the `musttail` attribute just introduced in https://reviews.llvm.org/D99517, we can now get guaranteed tail calls in both debug and opt builds, as long as we are building with a new enough version of Clang. This PR adopts the code base to take advantage of the `musttail` attribute when it is availabe.

This change unfortunately requires us to turn several `__attribute__((always_inline))` functions into macros to satisfy the constraint that caller and callee for `__attribute__((musttail))` have matching function signatures. Previously we were using these `always_inline` functions as pseudo-templates, with compile-time-known parameters functioning as pseudo template parameters. Unfortunately this strategy is incompatible with `musttail`'s constraints, so these pseudo-templates must become macros. The alternative would be to use C++ and real templates instead, and while the result would certainly be more readable, abandoning upb's pure-C nature would make other things more complex (for example, the amalgamation could no longer be a single `.c`/`.h` pair).

This change is basically code size neutral. There does appear to be a slight CPU regression, and looking at assembly output I can spot some obvious inefficiencies, like push/pop in our specialized parser functions. But it appears that some regression had occurred prior to this change, and regaining this perf should probably be a separate optimization effort. Nothing in this change should fundamentally compromise the performance opportunities of this design.

```
name                                      old time/op  new time/op  delta        
ArenaOneAlloc                             21.7ns ± 0%  21.5ns ± 0%  -0.54%  (p=0.000 n=12+11)
ArenaInitialBlockOneAlloc                 6.34ns ± 0%  6.33ns ± 0%    ~     (p=0.099 n=11+10)
LoadDescriptor_Upb                        43.2µs ± 1%  43.3µs ± 1%    ~     (p=0.347 n=12+12)
LoadAdsDescriptor_Upb                     2.50ms ± 0%  2.49ms ± 0%  -0.52%  (p=0.000 n=11+12)
LoadDescriptor_Proto2                      240µs ± 0%   241µs ± 0%  +0.53%  (p=0.000 n=12+12)
LoadAdsDescriptor_Proto2                  12.8ms ± 0%  12.7ms ± 0%  -1.06%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<UseArena,Copy>         4.85µs ± 0%  4.96µs ± 0%  +2.16%  (p=0.000 n=12+10)
Parse_Upb_FileDesc<UseArena,Alias>        3.96µs ± 0%  4.10µs ± 0%  +3.39%  (p=0.000 n=12+11)
Parse_Upb_FileDesc<InitBlock,Copy>        4.44µs ± 0%  4.57µs ± 0%  +2.92%  (p=0.000 n=9+12)
Parse_Upb_FileDesc<InitBlock,Alias>       3.61µs ± 0%  3.75µs ± 0%  +3.97%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,NoArena,Copy>       29.2µs ± 0%  29.4µs ± 0%  +0.50%  (p=0.000 n=12+12)
Parse_Proto2<FileDesc,UseArena,Copy>      20.6µs ± 2%  20.8µs ± 2%    ~     (p=0.347 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.5µs ± 0%  16.4µs ± 0%  -0.60%  (p=0.000 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.6µs ± 0%  16.4µs ± 1%  -1.10%  (p=0.039 n=12+12)
SerializeDescriptor_Proto2                5.33µs ± 1%  5.38µs ± 1%  +0.99%  (p=0.000 n=12+11)
SerializeDescriptor_Upb                   12.3µs ± 0%  12.5µs ± 0%  +1.35%  (p=0.000 n=12+12)
                                                                                                                                      
                                                                                                                                      
    FILE SIZE        VM SIZE                                                                                                          
 --------------  --------------                                                                                                       
   +18%    +144  [ = ]       0    [Unmapped]                                                                                          
  -1.2%      -4  -1.2%      -4    [section .text]                                                                                     
    [NEW]      +2  [NEW]      +2    fastdecode_isdonefallback                                                                         
    [DEL]      -6  [DEL]      -6    fastdecode_longstring_noutf8                                                                      
  -0.3%    -140  -0.3%    -140    upb/decode_fast.c                                                                                   
    +5.8%     +64  +6.0%     +64    upb_pom_1bt_max64b                                                                                
    +2.7%     +64  +2.7%     +64    upb_ppv8_2bt                                                                                      
    +2.7%     +32  +2.8%     +32    upb_psm_1bt_max256b                                                                               
    +2.8%     +32  +3.0%     +32    upb_psm_1bt_max64b                                                                                
    +2.8%     +32  +3.0%     +32    upb_psm_2bt_max64b                                                                                
    +4.0%     +24  +4.2%     +24    upb_psv8_1bt                                                                                      
    +2.0%     +16  +2.1%     +16    upb_prf4_2bt                                                                                      
    +1.3%     +16  +1.4%     +16    upb_prz8_2bt                                                                                      
    -1.6%      -8  -1.7%      -8    upb_cob_1bt                                                                                       
    -0.6%     -12  -0.7%     -12    [4 Others]                                                                                        
    -2.5%     -16  -2.6%     -16    upb_pov4_1bt                                                                                      
    -1.9%     -16  -2.0%     -16    upb_ppf8_2bt                                                                                      
    -1.3%     -16  -1.3%     -16    upb_prv8_2bt                                                                                      
    -2.5%     -16  -2.7%     -16    upb_psv4_1bt                                                                                      
    -2.5%     -16  -2.6%     -16    upb_psv4_2bt                                                                                      
    -3.0%     -32  -3.1%     -32    upb_prs_2bt                                                                                       
    -2.6%     -32  -2.6%     -32    upb_prv4_2bt                                                                                      
    -4.9%     -48  -5.1%     -48    upb_prb_2bt                                                                                       
    -3.9%     -48  -4.0%     -48    upb_prv4_1bt                                                                                      
    -7.2%     -72  -7.5%     -72    upb_prb_1bt                                                                                       
    -7.8%     -88  -8.0%     -88    upb_prs_1bt                                                                                       
  [ = ]       0  -0.1%    -144    TOTAL 
```